requirements.txt
streamlit
pandas
plotly
python-dotenv
google-analytics-data
google-auth
google-auth-oauthlib
